change current composition to deprecated
