+++ src/metatrain/utils/additive/__init__.py
@@ -1,4 +1,4 @@
-from .composition import CompositionModel  # noqa: F401
-from .composition_refactor import BaseCompositionModel  # noqa: F401
+from .composition import MetatrainCompositionModel  # noqa: F401
+from .deprecated_composition import CompositionModel  # noqa: F401
 from .remove import remove_additive  # noqa: F401
 from .zbl import ZBL  # noqa: F401
