+++ src/metatrain/utils/additive/_base_composition.py
@@ -1,13 +1,14 @@
 """
 Contains the ``BaseCompositionModel class. This is intended for eventual porting to
-metatensor-learn. The class ``CompositionModel`` wraps this to be compatible with
+metatomic. The class ``MetatrainCompositionModel`` wraps this to be compatible with
 metatrain-style objects.
 """
+
 from typing import Dict, List
-import torch
 
 import metatensor.torch as mts
-from metatensor.torch import TensorMap, TensorBlock, Labels, LabelsEntry
+import torch
+from metatensor.torch import Labels, LabelsEntry, TensorBlock, TensorMap
 from metatensor.torch.atomistic import System
 from metatensor.torch.learn.data import DataLoader
 
@@ -20,26 +21,29 @@ class BaseCompositionModel(torch.nn.Module):
         atomic_types: List[int],
         layouts: Dict[str, TensorMap],
     ) -> None:
-        
         target_names = []
         sample_kinds = {}
         for target_name, layout in layouts.items():  # identify target_type
-            
             target_names.append(target_name)
             if layout.sample_names == ["system"]:
                 sample_kinds[target_name] = "per_structure"
-            
+
             elif layout.sample_names == ["system", "atom"]:
                 sample_kinds[target_name] = "per_atom"
 
             elif layout.sample_names == [
-                "system", "first_atom", "second_atom", "cell_shift_a", "cell_shift_b", "cell_shift_c"
+                "system",
+                "first_atom",
+                "second_atom",
+                "cell_shift_a",
+                "cell_shift_b",
+                "cell_shift_c",
             ]:
                 sample_kinds[target_name] = "per_pair"
 
             else:
                 raise ValueError
-            
+
         self.atomic_types = atomic_types
         self.target_names = target_names
         self.sample_kinds = sample_kinds
@@ -48,13 +52,7 @@ class BaseCompositionModel(torch.nn.Module):
         in_keys = {
             target_name: Labels(
                 layout.keys.names,
-                torch.vstack(
-                    [
-                        key.values
-                        for key in layout.keys
-                        if _include_key(key)
-                    ]
-                )
+                torch.vstack([key.values for key in layout.keys if _include_key(key)]),
             )
             for target_name, layout in layouts.items()
         }
@@ -70,21 +68,26 @@ class BaseCompositionModel(torch.nn.Module):
                 blocks=[
                     TensorBlock(
                         values=torch.zeros(
-                            len(self.atomic_types), len(self.atomic_types),
+                            len(self.atomic_types),
+                            len(self.atomic_types),
                             dtype=torch.float64,
                         ),
                         samples=Labels(
-                            ["first_atom_type"],
-                            torch.tensor(self.atomic_types, dtype=torch.int32).reshape(-1, 1),
+                            ["center_type"],
+                            torch.tensor(self.atomic_types, dtype=torch.int32).reshape(
+                                -1, 1
+                            ),
                         ),
                         components=[],
                         properties=Labels(
-                            ["second_atom_type"],
-                            torch.tensor(self.atomic_types, dtype=torch.int32).reshape(-1, 1),
+                            ["center_type"],
+                            torch.tensor(self.atomic_types, dtype=torch.int32).reshape(
+                                -1, 1
+                            ),
                         ),
                     )
                     for _ in layout
-                ]
+                ],
             )
             for target_name, layout in layouts.items()
         }
@@ -101,16 +104,15 @@ class BaseCompositionModel(torch.nn.Module):
                         ),
                         samples=Labels(
                             ["center_type"],
-                            torch.tensor(
-                                self.atomic_types,
-                                dtype=torch.int32
-                            ).reshape(-1, 1),
+                            torch.tensor(self.atomic_types, dtype=torch.int32).reshape(
+                                -1, 1
+                            ),
                         ),
                         components=block.components,
                         properties=block.properties,
                     )
                     for block in layout
-                ]
+                ],
             )
             for target_name, layout in layouts.items()
         }
@@ -120,13 +122,11 @@ class BaseCompositionModel(torch.nn.Module):
         self,
         systems: List[System],
         targets: Dict[str, TensorMap],
-    ):
-        
+    ) -> None:
         num_atoms = torch.tensor([len(system) for system in systems])
-        
+
         for target_name, target in targets.items():
             for key, block in target.items():
-
                 if not _include_key(key):
                     continue
 
@@ -134,22 +134,21 @@ class BaseCompositionModel(torch.nn.Module):
                 Y = block.values
 
                 if self.sample_kinds[target_name] == "per_structure":
-
                     X = self._compute_X_per_structure(systems)
 
                     # For per-structure, divide target values by number of atoms
                     Y /= num_atoms.reshape(-1, *Y.shape[1:])
 
                 elif self.sample_kinds[target_name] in ["per_atom", "per_pair"]:
-
-                    X = self._compute_X_per_atom(systems, self._get_sliced_atomic_types(key))
+                    X = self._compute_X_per_atom(
+                        systems, self._get_sliced_atomic_types(key)
+                    )
 
                 else:
                     raise ValueError(
                         f"unknown sample kind: {self.sample_kinds[target_name]}"
                         f" for target {target_name}"
                     )
-                    
 
                 # Compute a sparse XTX
                 self.XTX[target_name][key].values[:] += X.T @ X
@@ -157,31 +156,31 @@ class BaseCompositionModel(torch.nn.Module):
                 # Compute XTY
                 self.XTY[target_name][key].values[:] += _compute_XTY(X, Y)
 
-
     def fit(self, dataloader: DataLoader) -> None:
+        """
+        Iterates over the batches in ``dataloader`` and accumulates the necessary
+        qunatities. Then, fits the compositions for each target.
+        """
 
         # acccumulate
         for batch in dataloader:
             self._accumulate(
                 batch.systems,
-                {
-                    target_name: batch[target_name]
-                    for target_name in self.target_names
-                },
+                {target_name: batch[target_name] for target_name in self.target_names},
             )
 
         # fit
         for target_name in self.target_names:
-
             blocks = []
             if self.sample_kinds[target_name] == "per_structure":
                 for key in self.XTX[target_name].keys:
-
                     XTX_block = self.XTX[target_name][key]
                     XTY_block = self.XTY[target_name][key]
                     blocks.append(
                         TensorBlock(
-                            values=_solve_linear_system(XTX_block.values, XTY_block.values),
+                            values=_solve_linear_system(
+                                XTX_block.values, XTY_block.values
+                            ),
                             samples=XTY_block.samples,
                             components=XTY_block.components,
                             properties=XTY_block.properties,
@@ -189,10 +188,8 @@ class BaseCompositionModel(torch.nn.Module):
                     )
 
             elif self.sample_kinds[target_name] in ["per_atom", "per_pair"]:
-
                 blocks = []
                 for key in self.XTX[target_name].keys:
-
                     XTX_block = self.XTX[target_name][key]
                     XTY_block = self.XTY[target_name][key]
 
@@ -220,19 +217,17 @@ class BaseCompositionModel(torch.nn.Module):
                     f"unknown sample kind: {self.sample_kinds[target_name]}"
                     f" for target {target_name}"
                 )
-            
+
             self.weights[target_name] = TensorMap(self.XTX[target_name].keys, blocks)
 
-
     def forward(self, systems: List[System]) -> Dict[str, TensorMap]:
+        """Predicts the compositions for the input ``systems``."""
 
         outputs = {}
         for target_name, weights in self.weights.items():
-            
             output_key_vals = []
             output_blocks = []
             for key, weight_block in weights.items():
-
                 # Compute X
                 if self.sample_kinds[target_name] == "per_structure":
                     X = self._compute_X_per_structure(systems)
@@ -247,7 +242,7 @@ class BaseCompositionModel(torch.nn.Module):
                         f"unknown sample kind: {self.sample_kinds[target_name]}"
                         f" for target {target_name}"
                     )
-                
+
                 # Predict this block
                 output_key_vals.append(key.values)
 
@@ -255,7 +250,9 @@ class BaseCompositionModel(torch.nn.Module):
                 output_blocks.append(
                     TensorBlock(
                         values=out_vals,
-                        samples=Labels(["_"], torch.arange(out_vals.shape[0]).reshape(-1, 1)),
+                        samples=Labels(
+                            ["_"], torch.arange(out_vals.shape[0]).reshape(-1, 1)
+                        ),
                         components=weight_block.components,
                         properties=weight_block.properties,
                     )
@@ -268,31 +265,25 @@ class BaseCompositionModel(torch.nn.Module):
                 ),
                 output_blocks,
             )
-        
-        return outputs
 
+        return outputs
 
     def _get_sliced_atomic_types(self, key) -> List[int]:
         """
         Gets the slice of atomic types needed for the block indexed by the input ``key``
         """
-        
+
         center_types = self.atomic_types
         if "center_type" in key.names:
             center_types = [key["center_type"]]
 
-        if (
-            "first_atom_type" in key.names
-            and "second_atom_type" in key.names
-        ):
+        if "first_atom_type" in key.names and "second_atom_type" in key.names:
             assert (
-                key["first_atom_type"] == key["second_atom_type"] 
-                and key["s2_pi"] == 0
+                key["first_atom_type"] == key["second_atom_type"] and key["s2_pi"] == 0
             )
             center_types = [key["first_atom_type"]]
 
         return center_types
-
 
     def _compute_X_per_structure(self, systems: List[System]) -> torch.Tensor:
         X = []
@@ -302,14 +293,15 @@ class BaseCompositionModel(torch.nn.Module):
                     torch.sum(system.types == atom_type)
                     for atom_type in self.atomic_types
                 ],
-                dtype=torch.float64
+                dtype=torch.float64,
             )
             X.append(X_system / len(system))
 
         return torch.vstack(X)
-    
-    def _compute_X_per_atom(self, systems: List[System], center_types: List[int]) -> torch.Tensor:
 
+    def _compute_X_per_atom(
+        self, systems: List[System], center_types: List[int]
+    ) -> torch.Tensor:
         # TODO: make this one hot encoding quicker
 
         column_idx_map = {atom_type: i for i, atom_type in enumerate(self.atomic_types)}
@@ -329,29 +321,24 @@ class BaseCompositionModel(torch.nn.Module):
 
 
 def _include_key(key: LabelsEntry) -> bool:
-
     include_key = False
 
     if len(key.names) == 1 and key.names[0] == "_":  # scalar
         include_key = True
 
     if "o3_lambda" in key.names:
-
         if key.names == ["o3_lambda", "o3_sigma"]:  # normal spherical target
-
             if key["o3_lambda"] == 0 and key["o3_sigma"] == 1:
                 include_key = True
 
         # For one-center targets include invariant blocks
         elif "center_type" in key.names:
-
             if key["o3_lambda"] == 0 and key["o3_sigma"] == 1:
                 include_key = True
 
         # For two-center targets include only invariant off-site blocks. Assumes
         # symmetrized for now
         elif "first_atom_type" in key.names and "second_atom_type" in key.names:
-
             if (
                 key["o3_lambda"] == 0
                 and key["o3_sigma"] == 1
@@ -366,35 +353,26 @@ def _include_key(key: LabelsEntry) -> bool:
     return include_key
 
 
-def _solve_linear_system(compf_t_at_compf, compf_t_at_targets) -> torch.Tensor:
-    trace_magnitude = float(torch.diag(compf_t_at_compf).abs().mean())
+def _solve_linear_system(
+    XTX_vals: torch.Tensor, XTY_vals: torch.Tensor
+) -> torch.Tensor:
+    trace_magnitude = float(torch.diag(XTX_vals).abs().mean())
     regularizer = 1e-14 * trace_magnitude
     return torch.linalg.solve(
-        compf_t_at_compf
+        XTX_vals
         + regularizer
         * torch.eye(
-            compf_t_at_compf.shape[1],
-            dtype=compf_t_at_compf.dtype,
-            device=compf_t_at_compf.device,
+            XTX_vals.shape[1],
+            dtype=XTX_vals.dtype,
+            device=XTX_vals.device,
         ),
-        compf_t_at_targets,
+        XTY_vals,
     )
 
+
 def _compute_XTY(X: torch.Tensor, Y: torch.Tensor) -> torch.Tensor:
-    if len(Y.shape) == 2:
-        idxs = "sZ,sP->ZP"
-    else:
-        assert len(Y.shape) == 3
-        idxs = "sZ,sCP->ZCP"
-    
-    return torch.einsum(idxs, X, Y)
+    return torch.tensordot(X, Y, dims=([0], [0]))
+
 
 def _compute_XTW(X: torch.Tensor, W: torch.Tensor) -> torch.Tensor:
-    
-    if len(W.shape) == 2:
-        idxs = "sZ,ZP->sP"
-    else:
-        assert len(W.shape) == 3
-        idxs = "sZ,ZCP->sCP"
-
-    return torch.einsum(idxs, X, W)
+    return torch.tensordot(X, W, dims=([1], [0]))
